fix: change init packages #99
