+++ packages/vvisp/referenceFiles/.gitignore
@@ -4,3 +4,4 @@ node_modules
 
 coverage.json
 package-lock.json
+.idea/
