--- packages/vvisp/referenceFiles/example.env
@@ -1,36 +1,0 @@
-# the name of network you want to connect, network on infura (required)
-# choose one of [local, mainnet, ropsten, kovan, rinkeby, custom]
-NETWORK= 'local'
-
-# configure your custom blockchain url (not technically required)
-URL=
-
-# the port number you want to connect (not technically required)
-PORT= 8545
-
-# configure your infura api key (not technically required)
-INFURA_API_KEY= '123123'
-
-# change the mnemonic that your hd wallet is seeded with (required)
-MNEMONIC= 'This should be your MNEMONIC word'
-
-# the derive path to MNEMONIC (not technically required, default is 0)
-PRIV_INDEX=
-
-# Private key for specific account. IT OVERRIDES MNEMONIC OPTIONS (not technically required)
-PRIVATE_KEY=
-
-# set amount of gas price you want (not technically required, default is 10000000000 wei)
-# Unit: wei (e.g. GAS_PRICE= 10000000000)
-GAS_PRICE=
-
-# set amount of gas limit you want (not technically required, default is 4600000 wei)
-# e.g. GAS_PRICE= 4600000
-GAS_LIMIT=
-
-# if you want to use other specific version of solc compiler, write version
-# if you don't need it, just keep it empty
-SOLC_VERSION=
-
-# write false if you don't want optimization, default is true
-SOLC_OPTIMIZATION=
