+++ packages/vvisp/referenceFiles/example.service.vvisp.json
@@ -5,19 +5,19 @@
   },
   "contracts": {
     "ContractKeyName1": {
-      "upgradeable": true,
-      "path": "path/to/your/contract/ContractV0.sol",
+      "path": "path/to/your/contract/Contract.sol",
+      "constructorArguments": [
+        "${contracts.ContractKeyName1.address}",
+        "${variables.varName}"
+      ],
       "initialize": {
         "functionName": "initialize",
         "arguments": ["argument1", "argument2"]
       }
     },
     "ContractKeyName2": {
-      "path": "path/to/your/contract/Contract.sol",
-      "constructorArguments": [
-        "${contracts.ContractKeyName1.address}",
-        "${variables.varName}"
-      ],
+      "upgradeable": true,
+      "path": "path/to/your/contract/ContractV0.sol",
       "initialize": {
         "functionName": "initialize",
         "arguments": ["argument1", "argument2"]
