+++ packages/vvisp/referenceFiles/package.json
@@ -7,16 +7,14 @@
     "test": "scripts/test.sh",
     "lint:sol": "solium -d .",
     "lint:sol:fix": "solium -d . --fix",
-    "coverage": "scripts/coverage.sh",
-    "compile": "truffle compile"
+    "coverage": "scripts/coverage.sh"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@haechi-labs/vvisp-utils": "^0.2.0",
-    "@haechi-labs/vvisp-contracts": "^0.1.1",
+    "@haechi-labs/vvisp-utils": "^1.0.0",
+    "@haechi-labs/vvisp-contracts": "^1.0.0",
     "coveralls": "^3.0.1",
-    "dotenv": "^5.0.1",
     "eth-gas-reporter": "^0.1.7",
     "ethereumjs-util": "^5.1.2",
     "ethjs-abi": "^0.2.1",
